backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
autogen-agentchat==0.5.3
autogen-core==0.5.3
autogen-ext==0.5.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
distro==1.9.0
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
jiter==0.9.0
jsonref==1.1.0
openai==1.75.0
opentelemetry-api==1.32.1
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
protobuf==5.29.4
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pytest==8.3.5
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
tavily-python==0.5.4
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
wrapt==1.17.2
zipp==3.21.0
